fix: 로그인 여부 확인 수정 & 프로필 수정 후 페이지 이동 삭제

+++ src/api/user.api.js
@@ -60,6 +60,7 @@ class UserAPI {
 
   async updateUserInfo(formData) {
     const accessToken = localStorage.getItem("accessToken");
+    console.log(formData);
     try {
       const path = "/profile";
       const response = await this.#axios.patch(path, formData, {
@@ -72,8 +73,13 @@ class UserAPI {
 
       return responseData;
     } catch (err) {
-      alert("AccessToken이 만료되었습니다!");
-      localStorage.clear();
+      if (!accessToken) {
+        alert("AccessToken이 만료되었습니다!");
+        localStorage.clear();
+      }
+      else{
+        alert(err.response?.message);
+      }
     }
   }
 }
